lending_back/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
google-api-python-client==2.95.0
google-auth==2.23.0
python-dotenv==1.0.0
requests==2.31.0
Pillow

